# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Leila API",
    "version": "1.0.0",
    "description": "AI-powered API for home service management with voice control capabilities",
    "contact": {
      "name": "API Support",
      "email": "api@leila.ai"
    }
  },
  "servers": [
    {
      "url": "https://api.leila.ai/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:5000/api/v1",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/bookings": {
      "post": {
        "summary": "Create a new booking",
        "tags": ["Bookings"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/services": {
      "get": {
        "summary": "Get available services",
        "tags": ["Services"],
        "responses": {
          "200": {
            "description": "List of services",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Service"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/webhooks": {
      "post": {
        "summary": "Register a webhook",
        "tags": ["Webhooks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook registered"
          }
        }
      }
    },
    "/chat": {
      "post": {
        "summary": "Send message to AI chat",
        "tags": ["Chat"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "BookingRequest": {
        "type": "object",
        "required": ["service", "customer", "appointment"],
        "properties": {
          "service": {
            "type": "string",
            "enum": ["plumbing", "electrical", "hvac", "cleaning", "lawn-care", "pest-control", "appliance-repair", "painting"]
          },
          "customer": {
            "type": "object",
            "required": ["firstName", "lastName", "email"],
            "properties": {
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "phone": {
                "type": "string"
              }
            }
          },
          "appointment": {
            "type": "object",
            "required": ["date", "time", "address"],
            "properties": {
              "date": {
                "type": "string",
                "format": "date"
              },
              "time": {
                "type": "string",
                "pattern": "^([0-1]?[0-9]|2[0-3]):[0-5][0-9]$"
              },
              "address": {
                "type": "string"
              },
              "notes": {
                "type": "string"
              }
            }
          }
        }
      },
      "BookingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "customer": {
            "type": "object"
          },
          "appointment": {
            "type": "object"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "priceRange": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          }
        }
      },
      "WebhookRequest": {
        "type": "object",
        "required": ["url", "events"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["booking.created", "booking.updated", "booking.cancelled", "technician.assigned", "job.completed"]
            }
          },
          "secret": {
            "type": "string",
            "description": "Secret for webhook signature validation"
          }
        }
      },
      "ChatMessage": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "context": {
            "type": "object"
          }
        }
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "response": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "suggestedActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}